Converting batch-ordinal from into to long.

+++ core/src/main/scala/org/apache/spark/sql/execution/columnar/encoding/UpdatedColumnDecoder.scala
@@ -31,11 +31,11 @@ import org.apache.spark.sql.types._
  * a nullable or non-nullable version as appropriate.
  */
 final class UpdatedColumnDecoder(decoder: ColumnDecoder, field: StructField,
-    delta1Position: Int, delta1: ColumnDeltaDecoder,
-    delta2Position: Int, delta2: ColumnDeltaDecoder,
-    delta3Position: Int, delta3: ColumnDeltaDecoder)
-    extends UpdatedColumnDecoderBase(decoder, field, delta1Position, delta1,
-      delta2Position, delta2, delta3Position, delta3) {
+    delta1Position: Long, delta1: ColumnDeltaDecoder,
+    delta2Position: Long, delta2: ColumnDeltaDecoder,
+    delta3Position: Long, delta3: ColumnDeltaDecoder)
+    extends UpdatedColumnDecoderBase(decoder, field,
+      delta1Position, delta1, delta2Position, delta2, delta3Position, delta3) {
 
   protected def nullable: Boolean = false
 
@@ -46,11 +46,11 @@ final class UpdatedColumnDecoder(decoder: ColumnDecoder, field: StructField,
  * Nullable version of [[UpdatedColumnDecoder]].
  */
 final class UpdatedColumnDecoderNullable(decoder: ColumnDecoder, field: StructField,
-    delta1Position: Int, delta1: ColumnDeltaDecoder,
-    delta2Position: Int, delta2: ColumnDeltaDecoder,
-    delta3Position: Int, delta3: ColumnDeltaDecoder)
-    extends UpdatedColumnDecoderBase(decoder, field, delta1Position, delta1,
-      delta2Position, delta2, delta3Position, delta3) {
+    delta1Position: Long, delta1: ColumnDeltaDecoder,
+    delta2Position: Long, delta2: ColumnDeltaDecoder,
+    delta3Position: Long, delta3: ColumnDeltaDecoder)
+    extends UpdatedColumnDecoderBase(decoder, field,
+      delta1Position, delta1, delta2Position, delta2, delta3Position, delta3) {
 
   protected def nullable: Boolean = true
 
@@ -65,21 +65,21 @@ object UpdatedColumnDecoder {
     // positions are initialized at max so that they always are greater
     // than a valid index
 
-    var delta1Position = Int.MaxValue
+    var delta1Position = Long.MaxValue
     val delta1 = if (delta1Buffer ne null) {
       val d = new ColumnDeltaDecoder(delta1Buffer, field)
       delta1Position = d.moveToNextPosition()
       d
     } else null
 
-    var delta2Position = Int.MaxValue
+    var delta2Position = Long.MaxValue
     val delta2 = if (delta2Buffer ne null) {
       val d = new ColumnDeltaDecoder(delta2Buffer, field)
       delta2Position = d.moveToNextPosition()
       d
     } else null
 
-    var delta3Position = Int.MaxValue
+    var delta3Position = Long.MaxValue
     val delta3 = if (delta3Buffer ne null) {
       val d = new ColumnDeltaDecoder(delta3Buffer, field)
       delta3Position = d.moveToNextPosition()
@@ -99,15 +99,15 @@ object UpdatedColumnDecoder {
 }
 
 abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructField,
-    private final var delta1Position: Int, delta1: ColumnDeltaDecoder,
-    private final var delta2Position: Int, delta2: ColumnDeltaDecoder,
-    private final var delta3Position: Int, delta3: ColumnDeltaDecoder) {
+    private final var delta1Position: Long, delta1: ColumnDeltaDecoder,
+    private final var delta2Position: Long, delta2: ColumnDeltaDecoder,
+    private final var delta3Position: Long, delta3: ColumnDeltaDecoder) {
 
   protected def nullable: Boolean
 
   protected final var nextDeltaBuffer: ColumnDeltaDecoder = _
   protected final var currentDeltaBuffer: ColumnDeltaDecoder = _
-  protected final var nextUpdatedPosition: Int = Int.MinValue
+  protected final var nextUpdatedPosition: Long = Long.MinValue
 
   final def getCurrentDeltaBuffer: ColumnDeltaDecoder = currentDeltaBuffer
 
@@ -116,7 +116,7 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
   }
 
   protected final def moveToNextUpdatedPosition1(ordinal: Int): Boolean = {
-    if (delta1Position.abs - 1 < Int.MaxValue && delta1Position.abs - 1 == ordinal) {
+    if (delta1Position.abs - 1 < Long.MaxValue && delta1Position.abs - 1 == ordinal) {
       nextUpdatedPosition = delta1Position
       delta1Position = delta1.moveToNextPosition()
       nextDeltaBuffer = delta1
@@ -126,7 +126,7 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
   }
 
   protected final def moveToNextUpdatedPosition2(ordinal: Int): Boolean = {
-    if (delta2Position.abs - 1 < Int.MaxValue && delta2Position.abs - 1== ordinal) {
+    if (delta2Position.abs - 1 < Long.MaxValue && delta2Position.abs - 1 == ordinal) {
       nextUpdatedPosition = delta2Position
       delta2Position = delta2.moveToNextPosition()
       nextDeltaBuffer = delta2
@@ -136,7 +136,7 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
   }
 
   protected final def moveToNextUpdatedPosition3(ordinal: Int): Boolean = {
-    if (delta3Position.abs - 1 < Int.MaxValue && delta3Position.abs - 1 == ordinal) {
+    if (delta3Position.abs - 1 < Long.MaxValue && delta3Position.abs - 1 == ordinal) {
       nextUpdatedPosition = delta3Position
       delta3Position = delta3.moveToNextPosition()
       nextDeltaBuffer = delta3
@@ -145,8 +145,8 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
     } else false
   }
 
-  private final def moveToNextUpdatedPosition(ordinal: Int): Int = {
-    var next = Int.MaxValue
+  private final def moveToNextUpdatedPosition(ordinal: Int): Long = {
+    var next = Long.MaxValue
     var movedIndex = -1
 
     // first delta is the lowest in hierarchy and overrides others
@@ -207,7 +207,7 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
         // skip the position in current delta
         skipUpdatedPosition(nextDeltaBuffer)
         // update the cursor and keep on till ordinal is not reached
-        nextUpdated = moveToNextUpdatedPosition(nextUpdated)
+        nextUpdated = moveToNextUpdatedPosition(nextUpdated.toInt)
       } while (nextUpdated < ordinal)
       nextUpdatedPosition = nextUpdated
       if (nextUpdated > ordinal) return true
@@ -219,15 +219,15 @@ abstract class UpdatedColumnDecoderBase(decoder: ColumnDecoder, field: StructFie
 
   final def unchanged(ordinal: Int, isCaseOfUpdate: Boolean): Boolean = {
     if (isCaseOfUpdate) {
-      if (nextUpdatedPosition == Int.MinValue) {
+      if (nextUpdatedPosition == Long.MinValue) {
         nextUpdatedPosition = moveToNextUpdatedPosition(-1)
       }
       // Original
       if (nextUpdatedPosition > ordinal) true
       else skipUntil(ordinal)
     } else {
-      if (nextUpdatedPosition == Int.MinValue) {
-        nextUpdatedPosition = Int.MaxValue
+      if (nextUpdatedPosition == Long.MinValue) {
+        nextUpdatedPosition = Long.MaxValue
       }
       if (moveToNextUpdatedPosition1(ordinal)) return false
       if (moveToNextUpdatedPosition2(ordinal)) return false
